fix: rollovers
