+++ server/tests/_groups/temp.ts
@@ -2,13 +2,13 @@ import type { TestGroup } from "./types";
 
 export const temp: TestGroup = {
 	name: "temp",
-	description: "Failed tests to triage and fix",
+	description: "Billing rollover regression suite (rollover carry-over fix)",
 	tier: "domain",
 	paths: [
-		"integration/balances/check/spend-limit/check-customer-spend-limit.test.ts",
-		"integration/balances/track/basic/track-event-name.test.ts",
-		"integration/balances/track/track-misc.test.ts",
-		"balances/track/entity-balances/track-entity-balances6.test.ts",
-		"balances/track/entity-balances/track-entity-balances7.test.ts",
+		"integration/billing/attach/immediate-switch/immediate-switch-rollover.test.ts",
+		"integration/billing/attach/scheduled-switch/scheduled-switch-rollover.test.ts",
+		"integration/billing/attach/scheduled-switch/discounts/scheduled-switch-discounts-edge.test.ts",
+		"integration/billing/create-schedule/create-schedule-basic.test.ts",
+		"integration/billing/update-subscription/custom-plan/update-paid-prepaid-rollover.test.ts",
 	],
 };
